First draft of generated data collection
With this feature we want to give the user a tool to understand which kind of data has been generated by the generators, so he/she will be able to tune them to maximize the efficiency of the quickcheck run on the given property.

This is a draft commit, many things need to be polished.